+++ pkg/providers/property/helpers.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
+	"regexp"
 	"strings"
 	"time"
 
@@ -13,6 +14,7 @@ import (
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v12/pkg/papi"
 	"github.com/akamai/terraform-provider-akamai/v9/pkg/common/str"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
 var certStatus = &schema.Resource{
@@ -170,6 +172,56 @@ var ccmCertificateStatusSchema = &schema.Resource{
 	},
 }
 
+var mtlsSchema = &schema.Resource{
+	Schema: map[string]*schema.Schema{
+		"ca_set_id": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "The ID of the Certificate Authority (CA) set to use for mTLS.",
+			ValidateDiagFunc: validation.ToDiagFunc(validation.StringMatch(regexp.MustCompile(`^\d+$`),
+				"must be a string representing an integer value")),
+		},
+		"check_client_ocsp": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Indicates whether to check the client OCSP.",
+		},
+		"send_ca_set_client": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Indicates whether to send the CA set to the client.",
+		},
+	},
+}
+
+var tlsConfigurationSchema = &schema.Resource{
+	Schema: map[string]*schema.Schema{
+		"cipher_profile": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "The cipher profile to use for TLS connections.",
+		},
+		"disallowed_tls_versions": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "A list of TLS versions that are disallowed.",
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
+		"staple_server_ocsp_response": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Indicates whether to staple the server OCSP response.",
+		},
+		"fips_mode": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Indicates whether FIPS mode is enabled.",
+		},
+	},
+}
+
 // Convert given hostnames to the map form that can be stored in a schema.ResourceData
 // Setting only statuses for default certs if they exist
 func flattenHostnames(Hostnames []papi.Hostname) []map[string]interface{} {
@@ -201,6 +253,8 @@ func flattenHostnamesCCM(Hostnames []papi.Hostname) []map[string]interface{} {
 		m["cert_status"] = []map[string]any{flattenCertType(&hn.CertStatus)}
 		m["ccm_certificates"] = flattenCCMCertificates(hn.CCMCertificates)
 		m["ccm_cert_status"] = flattenCCMCertificateStatus(hn.CCMCertStatus)
+		m["mtls"] = flattenMTLS(hn.MTLS)
+		m["tls_configuration"] = flattenTLSConfiguration(hn.TLSConfiguration)
 		res = append(res, m)
 	}
 	return res
@@ -227,6 +281,34 @@ func flattenCCMCertificates(certificates *papi.CCMCertificates) []map[string]str
 	m["rsa_cert_id"] = certificates.RSACertID
 	m["ecdsa_cert_id"] = certificates.ECDSACertID
 	return []map[string]string{m}
+}
+
+func flattenMTLS(mtls *papi.MTLS) []map[string]any {
+	if mtls == nil {
+		return nil
+	}
+	return []map[string]any{
+		{
+			"ca_set_id":          mtls.CASetID,
+			"check_client_ocsp":  mtls.CheckClientOCSP,
+			"send_ca_set_client": mtls.SendCASetClient,
+		},
+	}
+}
+
+func flattenTLSConfiguration(tlsConfig *papi.TLSConfiguration) []map[string]any {
+	if tlsConfig == nil {
+		return nil
+	}
+
+	return []map[string]any{
+		{
+			"cipher_profile":              tlsConfig.CipherProfile,
+			"disallowed_tls_versions":     tlsConfig.DisallowedTLSVersions,
+			"staple_server_ocsp_response": tlsConfig.StapleServerOcspResponse,
+			"fips_mode":                   tlsConfig.FIPSMode,
+		},
+	}
 }
 
 func flattenBucketHostnames(hostnames []papi.HostnameItem) []map[string]any {