+++ pkg/providers/property/resource_akamai_property.go
@@ -188,6 +188,20 @@ func resourceProperty() *schema.Resource {
 							Description: "Deployment status for the RSA and ECDSA certificates created with Cloud Certificate Manager (CCM).",
 							Elem:        ccmCertificateStatusSchema,
 						},
+						"mtls": {
+							Type:        schema.TypeList,
+							Optional:    true,
+							MaxItems:    1,
+							Description: "Optional mutual TLS settings for the CCM hostnames.",
+							Elem:        mtlsSchema,
+						},
+						"tls_configuration": {
+							Type:        schema.TypeList,
+							Optional:    true,
+							MaxItems:    1,
+							Description: "Optional TLS configuration settings applicable to the Cloud Certificate Manager (CCM) hostnames.",
+							Elem:        tlsConfigurationSchema,
+						},
 					},
 				},
 			},
@@ -305,7 +319,19 @@ func getCCMHashPart(hostname map[string]any) string {
 	}
 	rsaCertID := ccmCerts["rsa_cert_id"].(string)
 	ecdsaCertID := ccmCerts["ecdsa_cert_id"].(string)
-	return fmt.Sprintf(".%s.%s", rsaCertID, ecdsaCertID)
+	mtls := hostname["mtls"].([]any)
+	var caSetID string
+	if len(mtls) > 0 {
+		mtlsMap := mtls[0].(map[string]any)
+		caSetID = mtlsMap["ca_set_id"].(string)
+	}
+	tlsConfiguration := hostname["tls_configuration"].([]any)
+	var cipherProfile string
+	if len(tlsConfiguration) > 0 {
+		tlsConfigMap := tlsConfiguration[0].(map[string]any)
+		cipherProfile = tlsConfigMap["cipher_profile"].(string)
+	}
+	return fmt.Sprintf(".%s.%s.%s.%s", rsaCertID, ecdsaCertID, caSetID, cipherProfile)
 }
 
 // propertyRulesCustomDiff compares Rules.Criteria and Rules.Children fields from terraform state
@@ -507,6 +533,15 @@ func ensureCCMCertificatesConsistency(_ context.Context, d *schema.ResourceDiff,
 		} else {
 			if areCcmCerts {
 				return fmt.Errorf("ccm_certificates is only allowed when cert_provisioning_type is 'CCM'")
+			}
+			if len(m["mtls"].([]any)) > 0 {
+				return fmt.Errorf("hostname %v: mtls can only be set when cert_provisioning_type is CCM",
+					m["cname_from"])
+			}
+			if len(m["tls_configuration"].([]any)) > 0 {
+				return fmt.Errorf(
+					"hostname %v: tls_configuration can only be set when cert_provisioning_type is CCM",
+					m["cname_from"])
 			}
 		}
 	}
@@ -1333,7 +1368,7 @@ func fetchPropertyVersion(ctx context.Context, client papi.PAPI, propertyID, gro
 	return res, err
 }
 
-// fetchPropertyVersionHostnames fetchs hostnames for latest version of given property.
+// fetchPropertyVersionHostnames fetches hostnames for latest version of given property.
 func fetchPropertyVersionHostnames(ctx context.Context, client papi.PAPI, property papi.Property, version int) ([]papi.Hostname, error) {
 	req := papi.GetPropertyVersionHostnamesRequest{
 		PropertyID:        property.PropertyID,
@@ -1517,6 +1552,8 @@ func mapToHostnames(givenList []interface{}) []papi.Hostname {
 		cnameTo := r["cname_to"]
 		certProvisioningType := r["cert_provisioning_type"]
 		if len(r) != 0 {
+			var mtls *papi.MTLS
+			var tlsConfig *papi.TLSConfiguration
 			var ccmCerts *papi.CCMCertificates
 			if certProvisioningType.(string) == string(papi.CertTypeCCM) {
 				certs := r["ccm_certificates"].([]any)
@@ -1527,6 +1564,36 @@ func mapToHostnames(givenList []interface{}) []papi.Hostname {
 						ECDSACertID: m["ecdsa_cert_id"].(string),
 					}
 				}
+
+				if r["mtls"] != nil {
+					mtlsMap := r["mtls"].([]any)
+					if len(mtlsMap) > 0 {
+						m := mtlsMap[0].(map[string]any)
+						mtls = &papi.MTLS{
+							CASetID:         m["ca_set_id"].(string),
+							CheckClientOCSP: m["check_client_ocsp"].(bool),
+							SendCASetClient: m["send_ca_set_client"].(bool),
+						}
+					}
+				}
+
+				if r["tls_configuration"] != nil {
+					tlsConfigMap := r["tls_configuration"].([]any)
+					if len(tlsConfigMap) > 0 {
+						m := tlsConfigMap[0].(map[string]any)
+
+						var disallowedTLSVersions []string
+						for _, v := range m["disallowed_tls_versions"].([]any) {
+							disallowedTLSVersions = append(disallowedTLSVersions, v.(string))
+						}
+						tlsConfig = &papi.TLSConfiguration{
+							CipherProfile:            m["cipher_profile"].(string),
+							DisallowedTLSVersions:    disallowedTLSVersions,
+							StapleServerOcspResponse: m["staple_server_ocsp_response"].(bool),
+							FIPSMode:                 m["fips_mode"].(bool),
+						}
+					}
+				}
 			}
 
 			hostnames = append(hostnames, papi.Hostname{
@@ -1535,8 +1602,9 @@ func mapToHostnames(givenList []interface{}) []papi.Hostname {
 				CnameTo:              cnameTo.(string), // guaranteed by schema to be a string
 				CertProvisioningType: certProvisioningType.(string),
 				CCMCertificates:      ccmCerts,
+				MTLS:                 mtls,
+				TLSConfiguration:     tlsConfig,
 			})
-
 		}
 	}
 	return hostnames
@@ -1569,7 +1637,6 @@ func validateAtLeastOneCertIDProvidedCCM(d *schema.ResourceData) error {
 					r["cname_from"])
 			}
 		}
-
 	}
 	return nil
 }