+++ pkg/providers/property/testdata/TestResProperty/CCM/update_mtls_and_tls_configuration.tf
@@ -1,0 +1,26 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_property" "test" {
+  name        = "test_property"
+  contract_id = "ctr_C-0N7RAC7"
+  group_id    = "grp_12345"
+  product_id  = "prd_Object_Delivery"
+
+  hostnames {
+    cname_from             = "example.com"
+    cname_to               = "example.com.edgekey.net"
+    cert_provisioning_type = "CCM"
+    ccm_certificates {
+      rsa_cert_id = "654321"
+    }
+    mtls {
+      ca_set_id = "524126"
+    }
+    tls_configuration {
+      cipher_profile          = "ak-akamai-2020q2"
+      disallowed_tls_versions = ["TLSv1", "TLSv2"]
+    }
+  }
+}